requirements.txt
appdirs==1.4.4
asgiref==3.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
ChatterBot==1.0.0
chatterbot-corpus @ git+https://github.com/gunthercox/chatterbot-corpus/@ae8ccd2912baa79cf0412c3c0822835dab64059c
chirptext==0.1.2
click==8.1.3
colorama==0.4.5
comtypes==1.4.7
Deprecated==1.2.14
distlib==0.3.6
Django==4.1.1
filelock==3.8.0
Flask==2.2.2
Flask-MySQLdb==1.0.1
flexcache==0.3
flexparser==0.3.1
fsspec==2024.9.0
fugashi==1.3.2
idna==3.6
itsdangerous==2.1.2
jaconv==0.4.0
jamdict==0.1a11.post2
Jinja2==3.1.2
joblib==1.4.2
libretranslatepy==2.1.1
llvmlite==0.43.0
lxml==5.1.0
MarkupSafe==2.1.1
mathparse==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
mysql-connector-python==8.0.30
mysqlclient==2.1.1
networkx==3.3
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20231117
pandas==2.2.2
Pillow==9.2.0
Pint==0.24.3
platformdirs==2.5.2
protobuf==3.20.1
puchikarui==0.1
pydub==0.25.1
pygame==2.6.0
pykakasi==2.3.0
pymongo==3.13.0
PyMySQL==1.0.2
pypandoc==1.5
pypiwin32==223
python-dateutil==2.7.5
pyttsx3==2.91
pytz==2024.1
pywin32==306
PyYAML==3.13
regex==2024.7.24
requests==2.31.0
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
SpeechRecognition==3.10.4
SQLAlchemy==1.2.19
sqlparse==0.4.3
sympy==1.13.2
threadpoolctl==3.5.0
tiktoken==0.7.0
torch==2.4.1
tqdm==4.66.5
translate==3.6.1
typing_extensions==4.12.2
tzdata==2024.1
unidic-lite==1.0.8
urllib3==2.2.1
virtualenv==20.16.5
virtualenvwrapper-win==1.2.7
Werkzeug==2.2.2
wrapt==1.16.0
